Fix code to meet standards set and remove duplicated code
