+++ app/src/androidTest/java/com/andela/art/login/LoginActivityTest.java
@@ -4,7 +4,6 @@ package com.andela.art.login;
 import android.support.test.espresso.ViewInteraction;
 import android.support.test.rule.ActivityTestRule;
 import android.support.test.runner.AndroidJUnit4;
-import android.test.suitebuilder.annotation.LargeTest;
 import android.view.View;
 import android.view.ViewGroup;
 import android.view.ViewParent;
@@ -14,7 +13,6 @@ import com.andela.art.R;
 import org.hamcrest.Description;
 import org.hamcrest.Matcher;
 import org.hamcrest.TypeSafeMatcher;
-import org.junit.Ignore;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -29,8 +27,6 @@ import static org.hamcrest.Matchers.allOf;
  * Tests for the login screen, the main screen which contains a user is presented on first launch
  * of the application.
  */
-@Ignore
-@LargeTest
 @RunWith(AndroidJUnit4.class)
 public class LoginActivityTest {
 
@@ -62,6 +58,7 @@ public class LoginActivityTest {
                                 1),
                         isDisplayed()));
 //        googleSignInButton.perform(click());
+
     }
 
     /**
