+++ app/src/androidTest/java/com/andela/art/root/AppTest.java
@@ -9,7 +9,6 @@ import android.support.test.uiautomator.UiDevice;
 import com.andela.art.login.LoginActivity;
 
 import org.junit.Assert;
-import org.junit.Ignore;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -20,7 +19,6 @@ import static org.junit.Assert.assertEquals;
 /**
  * Created by Jeffkungu on 14/03/2018.
  */
-@Ignore
 @RunWith(AndroidJUnit4.class)
 public class AppTest {
