+++ app/src/androidTest/java/com/andela/art/settings/SettingsTest.java
@@ -10,7 +10,6 @@ import com.andela.art.login.LoginActivity;
 
 import org.junit.After;
 import org.junit.Before;
-import org.junit.Ignore;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -30,7 +29,6 @@ import static android.support.test.espresso.matcher.ViewMatchers.withText;
 /**
  * Tests for the settings activity.
  */
-@Ignore
 @RunWith(AndroidJUnit4.class)
 public class SettingsTest {
