+++ app/src/main/AndroidManifest.xml
@@ -6,10 +6,8 @@
 
     <uses-permission android:name="android.permission.INTERNET" />
 
-    <uses-permission android:name="android.permission.INTERNET" />
-
     <application
-        android:name=".common.ArtApplication"
+        android:name=".root.ArtApplication"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
