--- app/src/main/java/com/andela/art/login/LoginModel.java
@@ -1,18 +1,0 @@
-package com.andela.art.login;
-
-/**
- * Created by Mugiwara_Munyi on 28/02/2018.
- */
-
-public class LoginModel implements LoginActivityMVP.LoginModel {
-    private final LoginRepository repository;
-
-    /**
-     * Perform database operations.
-     *
-     * @param repo - repository.
-     */
-    public LoginModel(LoginRepository repo) {
-        this.repository = repo;
-    }
-}
