+++ app/src/main/java/com/andela/art/login/injection/LoginModule.java
@@ -1,4 +1,4 @@
-package com.andela.art.login;
+package com.andela.art.login.injection;
 
 import android.content.Context;
 
@@ -16,40 +16,6 @@ import dagger.Provides;
 
 @Module
 public class LoginModule {
-
-    /**
-     * Provide the loginPresenter.
-     *
-     * @param model - LoginModel
-     * @return LoginPresenter class
-     */
-    @Provides
-    public LoginActivityMVP.LoginPresenter provideLoginActivityPresenter(
-            LoginModel model) {
-        return new LoginPresenter(model);
-    }
-
-    /**
-     * Provide loginModel.
-     *
-     * @param repository - LoginRepository
-     * @return LoginModel class
-     */
-    @Provides
-    public LoginActivityMVP.LoginModel provideLoginActivityModel(LoginRepository repository) {
-        return new LoginModel(repository);
-    }
-
-    /**
-     * Provide LoginRepository.
-     *
-     * @return LoginRepository class
-     */
-    @Provides
-    public LoginRepository provideLoginRepository() {
-        return new LoginRepository();
-    }
-
     /**
      * Provides GoogleSignInClient.
      *
