+++ app/src/main/java/com/andela/art/root/Activity.java
@@ -1,4 +1,4 @@
-package com.andela.art.common;
+package com.andela.art.root;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
