--- app/src/main/java/com/andela/art/root/App.java
@@ -1,32 +1,0 @@
-package com.andela.art.root;
-
-import android.app.Application;
-
-import com.andela.art.login.LoginModule;
-
-/**
- * Created by Mugiwara_Munyi on 28/02/2018.
- */
-
-public class App extends Application {
-
-    private ApplicationComponent component;
-
-    @Override
-    public void onCreate() {
-        super.onCreate();
-        component = DaggerApplicationComponent.builder()
-                .applicationModule(new ApplicationModule(this))
-                .loginModule(new LoginModule())
-                .build();
-    }
-
-    /**
-     * Get component.
-     *
-     * @return component - return the specified component.
-     */
-    public ApplicationComponent getComponent() {
-        return component;
-    }
-}
