+++ app/src/main/java/com/andela/art/root/Presenter.java
@@ -1,4 +1,4 @@
-package com.andela.art.common;
+package com.andela.art.root;
 
 
 /**
