+++ app/src/main/java/com/andela/art/securitydashboard/injection/SerialEntryComponent.java
@@ -1,7 +1,7 @@
 package com.andela.art.securitydashboard.injection;
 
 import com.andela.art.api.ApiModule;
-import com.andela.art.common.ApplicationComponent;
+import com.andela.art.root.ApplicationComponent;
 import com.andela.art.firebase.FirebaseModule;
 import com.andela.art.securitydashboard.presentation.SecurityDashboardActivity;
