+++ app/src/main/java/com/andela/art/securitydashboard/presentation/SecurityDashboardActivity.java
@@ -14,8 +14,8 @@ import android.widget.Toast;
 
 import com.andela.art.R;
 import com.andela.art.checkin.CheckInActivity;
-import com.andela.art.common.ApplicationComponent;
-import com.andela.art.common.ArtApplication;
+import com.andela.art.root.ApplicationComponent;
+import com.andela.art.root.ArtApplication;
 import com.andela.art.databinding.SecurityDashboardBinding;
 import com.andela.art.login.LoginActivity;
 import com.andela.art.securitydashboard.data.Asset;
