+++ app/src/prod/java/com/andela/art/securitydashboard/presentation/FirebasePresenter.java
@@ -3,7 +3,7 @@ package com.andela.art.securitydashboard.presentation;
 
 
 
-import com.andela.art.common.Presenter;
+import com.andela.art.root.Presenter;
 import com.google.firebase.auth.FirebaseAuth;
 
 /**
