--- app/src/test/java/com/andela/art/ExampleUnitTest.java
@@ -1,22 +1,0 @@
-package com.andela.art;
-
-import org.junit.Test;
-
-import static org.junit.Assert.assertEquals;
-
-/**
- * Example local unit test, which will execute on the development machine (host).
- *
- * @see <a href="http://d.android.com/tools/testing">Testing documentation</a>
- */
-public class ExampleUnitTest {
-    /**
-     * Test if addiiion of two numbers yields the corrent result.
-     *
-     * @throws Exception if an error occurs
-     */
-    @Test
-    public void addition_isCorrect() throws Exception {
-        assertEquals(4, 2 + 2);
-    }
-}
